+++ lib/typings/src/types/descriptor.ts
@@ -1,0 +1,3 @@
+import { Any } from './any';
+
+export type Descriptor = TypedPropertyDescriptor<Any>;