+++ lib/typings/src/types/dispatched-result.ts
@@ -1,0 +1,5 @@
+import { Observable } from 'rxjs';
+
+import { Any } from './any';
+
+export type DispatchedResult = Any | Observable<Any> | null;